pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "pytQt-template"
version = "2.0.0"
description = "PyQt Windows应用程序开发模板"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "PyQt Template Team"},
]
keywords = ["PyQt", "Qt", "GUI", "Windows", "Template"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "PySide6>=6.5.0",
    "loguru>=0.7.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "debugpy>=1.6.0",
    "ruff>=0.1.0",
    "basedpyright>=0.1.0",
]
epic = [
    "anthropic>=0.40.0",
    "claude-agent-sdk>=0.1.0",
    "pytest-qt>=4.2.0",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.11.0",
    "nuitka>=1.8.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing --json-report --json-report-file=test_results.json"
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "performance: marks tests as performance tests",
    "documentation: marks tests as documentation tests",
    "epic: marks tests as epic-related tests",
    "gui: marks tests as GUI tests",
    "acceptance: marks tests as acceptance criteria tests",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B018", "B017"]

[project.scripts]
bubble-sort = "src.cli:main"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
